requirements.txt
flask==3.1.1
flask-cors==5.0.1
gunicorn==23.0.0
anthropic==0.52.0
requests==2.32.3
google-api-python-client==2.166.0
google-auth==2.38.0
